Moves pattern_machine and sequence_machine files into generators module.
