+++ nupic/data/generators/pattern_machine.py
