+++ nupic/data/generators/sequence_machine.py
