+++ tests/integration/nupic/algorithms/extensive_temporal_memory_test.py
@@ -20,9 +20,9 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-import unittest2 as unittest
+import unittest
 
-from nupic.data.pattern_machine import PatternMachine
+from nupic.data.generators.pattern_machine import PatternMachine
 
 from nupic.test.abstract_temporal_memory_test import AbstractTemporalMemoryTest
