+++ tests/integration/nupic/algorithms/monitor_mixin/temporal_memory_monitor_mixin_test.py
@@ -22,8 +22,8 @@
 
 import unittest
 
-from nupic.data.pattern_machine import ConsecutivePatternMachine
-from nupic.data.sequence_machine import SequenceMachine
+from nupic.data.generators.pattern_machine import ConsecutivePatternMachine
+from nupic.data.generators.sequence_machine import SequenceMachine
 from nupic.research.temporal_memory import TemporalMemory
 from nupic.research.monitor_mixin.temporal_memory_monitor_mixin import (
   TemporalMemoryMonitorMixin)
