+++ tests/integration/nupic/algorithms/temporal_memory_performance_test.py
@@ -25,8 +25,8 @@ import unittest
 
 import numpy
 
-from nupic.data.pattern_machine import PatternMachine
-from nupic.data.sequence_machine import SequenceMachine
+from nupic.data.generators.pattern_machine import PatternMachine
+from nupic.data.generators.sequence_machine import SequenceMachine
 from nupic.research.temporal_memory import TemporalMemory
 from nupic.research.TP import TP
 from nupic.research.TP10X2 import TP10X2
