+++ tests/unit/nupic/data/pattern_machine_test.py
@@ -21,8 +21,8 @@
 # ----------------------------------------------------------------------
 import unittest
 
-from nupic.data.pattern_machine import (PatternMachine,
-                                        ConsecutivePatternMachine)
+from nupic.data.generators.pattern_machine import (PatternMachine,
+                                                   ConsecutivePatternMachine)
